% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exprB_example}
\alias{exprB_example}
\title{Example bulk expression matrix}
\format{
A numeric matrix with genes in rows and samples in columns.
}
\source{
Included with the SMiXcanK package.
}
\usage{
exprB_example
}
\description{
A small toy bulk expression matrix (genes x samples) used in the README.
}
\keyword{datasets}
