% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas_example}
\alias{gwas_example}
\title{Example GWAS summary statistics}
\format{
A list with elements \code{Beta} and \code{se_Beta}.
}
\source{
Included with the SMiXcanK package.
}
\usage{
gwas_example
}
\description{
A minimal list containing GWAS effect sizes and standard errors for the SNPs
in \code{x_example}.
}
\keyword{datasets}
