% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_example}
\alias{markers_example}
\title{Example marker gene list}
\format{
A named list of character vectors.
}
\source{
Included with the SMiXcanK package.
}
\usage{
markers_example
}
\description{
Marker genes for each cell type used by \code{pi_estimation_K()}.
}
\keyword{datasets}
