% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MiXcan_updated.R
\name{MiXcan_assoc_test}
\alias{MiXcan_assoc_test}
\title{Cell-Type-Aware Association with Shrinkage}
\usage{
MiXcan_assoc_test(
  outcome,
  cell1,
  cell2,
  family = "binomial",
  rho_thr = 0.999999
)
}
\arguments{
\item{outcome}{A binary phenotype vector (0/1 or factor).}

\item{cell1}{y predicted for cell type1.}

\item{cell2}{y predicted for cell type2.}

\item{family}{'binomial' or 'gaussian'}

\item{rho_thr}{threshold where choose sep model}
}
\value{
A data frame containing effect size estimates, standard errors, and p-values for both cell types,
and a combined p-value from ACAT.
}
\description{
This function performs cell-type-aware association analysis with penalization. It estimates the individual and combined p-values for
predicted gene expressions (e.g., from cell type 1 and cell type 2).
}
