% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S-MiXcan_K.R
\name{regularized_inverse_cov}
\alias{regularized_inverse_cov}
\title{Regularized Covariance Matrix Inversion}
\usage{
regularized_inverse_cov(X)
}
\arguments{
\item{X}{A numeric covariance matrix.}
}
\value{
A list containing:
\describe{
\item{inv}{The regularized inverse matrix.}
\item{lambda}{The regularization parameter used.}
}
}
\description{
Regularized Inverse of a Covariance Matrix
}
\details{
Computes a regularized inverse of a covariance matrix using correlation shrinkage.
}
