% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S-MiXcan_K.R
\name{SMiXcan_assoc_test_K}
\alias{SMiXcan_assoc_test_K}
\title{S-MiXcan Association Test in K cell types}
\usage{
SMiXcan_assoc_test_K(
  W,
  gwas_results,
  x_g,
  n0,
  n1,
  family = c("binomial", "gaussian")
)
}
\arguments{
\item{W}{A p-by-k matrix of cell-type level weights  (where p is the number of SNPs in the gene region, k is the number of cell types).}

\item{gwas_results}{A list containing GWAS summary statistics, with components \code{Beta} and \code{se_Beta}.}

\item{x_g}{A genotype matrix for the reference panel (individuals * SNPs).}

\item{n0}{Number of controls.}

\item{n1}{Number of cases.}

\item{family}{Either \code{"binomial"} or \code{"gaussian"} (used for fitting the null model).}
}
\value{
A list containing:
\describe{
\item{Z_join}{Z-score for cell-type 1 to K (joint model)}
\item{p_join_vec}{P-value for cell-type 1 to K (joint model)}
\item{p_join}{Combined p-value from joint model (ACAT)}
}
}
\description{
S-MiXcan Association Test with Shrinkage
}
\details{
Performs the S-MiXcan association test using GWAS summary statistics and reference genotype data,
applying shrinkage-based regularization to stabilize inverse covariance estimation.
}
