% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{y_example}
\alias{y_example}
\title{Example gene expression vector}
\format{
A numeric vector.
}
\source{
Included with the SMiXcanK package.
}
\usage{
y_example
}
\description{
A small toy expression vector for one gene (length = samples).
}
\keyword{datasets}
