% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genome_wide.R
\name{run_gwas}
\alias{run_gwas}
\title{Run Genome-Wide Association Study}
\usage{
run_gwas(X, D, family0)
}
\arguments{
\item{X}{A genotype matrix (individuals × SNPs).}

\item{D}{A phenotype vector (numeric or binary).}

\item{family0}{A GLM family, either "binomial" or "gaussian".}
}
\value{
A list with:
\describe{
\item{Beta}{Vector of effect sizes for each SNP}
\item{se_Beta}{Vector of standard errors for each SNP}
}
}
\description{
Run GWAS for Each SNP
}
\details{
Performs univariate GWAS by regressing the phenotype on each SNP separately.
}
